requirements.txt
# Core Flask dependencies
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
wsproto==1.2.0

# Database migrations
Flask-Migrate==4.0.5
alembic==1.13.1

# Email functionality (for future email verification/reset)
Flask-Mail==0.9.1

# Security enhancements
Flask-WTF==1.2.1
WTForms==3.1.1

# Rate limiting
Flask-Limiter==3.5.0
limits==3.6.0

# Environment variables
python-dotenv==1.0.0

# Password strength validation
password-strength==0.0.3.post2

# JWT tokens (for future API authentication)
PyJWT==2.8.0

# Google OAuth authentication

# Production server
gunicorn==21.2.0
Authlib==1.2.1
requests==2.31.0

# Additional security
cryptography==41.0.7
bcrypt==4.1.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
